feat: move top commands to its own module
