+++ CHANGELOG
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/lang/pt-BR/
 
 - O comando `/perfil` está **MUITO** mais rápido! Tanto na primeira execução, quanto na segunda (na segunda está extremamente rápido, quase instantâneo)
 - O comando `/top comandos tipo:comandos mais usados de um usuário` está **MUITO** mais rápido! É praticamente instantâneo.
+- A página máxima do top foi reduzida de `100` para `99`.
 
 ## [5.9.4-HALLOWEEN] - 2023-10-28
