+++ packages/events/src/modules/top/index.ts
@@ -1,0 +1,22 @@
+import { EMOJIS } from '../../structures/constants';
+
+const calculateSkipCount = (page: number): number => (page - 1) * 10;
+
+const topEmojis: { [key: string]: string } = {
+  mamou: EMOJIS.crown,
+  mamado: EMOJIS.lick,
+  estrelinhas: EMOJIS.estrelinhas,
+  demons: EMOJIS.demons,
+  giants: EMOJIS.giants,
+  angels: EMOJIS.angels,
+  archangels: EMOJIS.archangels,
+  demigods: EMOJIS.demigods,
+  gods: EMOJIS.gods,
+  votes: EMOJIS.ok,
+  blackjack: '🃏',
+  coinflip: '📀',
+  roulette: '🎡',
+  bicho: '🦌',
+};
+
+export { calculateSkipCount, topEmojis };
